traffic marker publisher class filled
