+++ simulation/traffic_simulator/include/traffic_simulator/traffic_lights/configurable_rate_updater.hpp
@@ -30,30 +30,14 @@ class ConfigurableRateUpdater {
 
   double publish_rate_ = 0.0;
 
-  const rclcpp::Publisher<visualization_msgs::msg::MarkerArray>::SharedPtr marker_pub_;
-
-  const std::string map_frame_;
-
-  auto deleteAllMarkers() const -> void;
-
-  auto drawMarkers() const -> void;
-
-  virtual auto publishTrafficLightStateArray() const -> void = 0;
-
 protected:
-  const std::shared_ptr<TrafficLightManager> traffic_light_manager_;
-
   const rclcpp::Clock::SharedPtr clock_ptr_;
 
 public:
   template <typename NodePointer>
-  ConfigurableRateUpdater(const std::shared_ptr<TrafficLightManager> & traffic_light_manager, const NodePointer & node, const std::string & map_frame = "map")
+  ConfigurableRateUpdater(const NodePointer & node)
   : node_base_interface_(node->get_node_base_interface())
   , node_timers_interface_(node->get_node_timers_interface())
-  , marker_pub_(rclcpp::create_publisher<visualization_msgs::msg::MarkerArray>(
-      node, "traffic_light/marker", rclcpp::QoS(1).transient_local()))
-  , map_frame_(map_frame)
-  , traffic_light_manager_(std::move(traffic_light_manager))
   , clock_ptr_(node->get_clock())
   {
 
@@ -63,7 +47,7 @@ public:
 
   auto resetPublishRate(double update_rate) -> void;
 
-  auto update(const double) -> void;
+  virtual auto update() -> void = 0;
 };
 }
