+++ simulation/traffic_simulator/src/traffic_lights/configurable_rate_updater.cpp
@@ -16,42 +16,6 @@
 
 namespace traffic_simulator
 {
-  auto ConfigurableRateUpdater::deleteAllMarkers() const -> void
-  {
-    visualization_msgs::msg::MarkerArray message;
-    {
-      visualization_msgs::msg::Marker marker;
-      marker.action = marker.DELETEALL;
-      message.markers.push_back(marker);
-    }
-
-    marker_pub_->publish(message);
-  }
-
-  auto ConfigurableRateUpdater::drawMarkers() const -> void
-  {
-    visualization_msgs::msg::MarkerArray marker_array;
-
-    const auto now = clock_ptr_->now();
-
-    for (const auto & [id, traffic_light] : traffic_light_manager_->getTrafficLights()) {
-      traffic_light.draw(marker_array.markers, now, map_frame_);
-    }
-
-    marker_pub_->publish(marker_array);
-  }
-
-  auto ConfigurableRateUpdater::update(const double) -> void
-  {
-    publishTrafficLightStateArray();
-
-    if (traffic_light_manager_->hasAnyLightChanged()) {
-      deleteAllMarkers();
-    }
-
-    drawMarkers();
-  }
-
   auto ConfigurableRateUpdater::createTimer(double publish_rate) -> void
   {
     if (!timer_) {
@@ -59,7 +23,7 @@ namespace traffic_simulator
       using namespace std::chrono_literals;
       timer_ = rclcpp::create_timer(
           node_base_interface_, node_timers_interface_, clock_ptr_, 1s / publish_rate_,
-          [this]() -> void { update(1.0 / publish_rate_); });
+          [this]() -> void { update(); });
     }
   }
 
@@ -74,7 +38,7 @@ namespace traffic_simulator
       using namespace std::chrono_literals;
       timer_ = rclcpp::create_timer(
           node_base_interface_, node_timers_interface_, clock_ptr_, 1s / publish_rate_,
-          [this]() -> void { update(1.0 / publish_rate_); });
+          [this]() -> void { update(); });
     }
   }
 }
