+++ simulation/traffic_simulator/src/traffic_lights/traffic_light_marker_publisher.cpp
@@ -14,3 +14,41 @@
 
 #include <traffic_simulator/traffic_lights/traffic_light_marker_publisher.hpp>
 
+namespace traffic_simulator {
+  auto TrafficLightMarkerPublisher::deleteAllMarkers() const -> void
+  {
+    visualization_msgs::msg::MarkerArray message;
+    {
+      visualization_msgs::msg::Marker marker;
+      marker.action = marker.DELETEALL;
+      message.markers.push_back(marker);
+    }
+
+    marker_pub_->publish(message);
+  }
+
+  auto TrafficLightMarkerPublisher::drawMarkers() const -> void
+  {
+    visualization_msgs::msg::MarkerArray marker_array;
+
+    const auto now = clock_ptr_->now();
+
+    for (const auto & [id, traffic_light] : traffic_light_manager_->getTrafficLights()) {
+      traffic_light.draw(marker_array.markers, now, map_frame_);
+    }
+
+    marker_pub_->publish(marker_array);
+  }
+
+  auto TrafficLightMarkerPublisher::update() -> void
+  {
+    if (traffic_light_manager_->hasAnyLightChanged()) {
+      deleteAllMarkers();
+    }
+
+    drawMarkers();
+  }
+
+
+}
+
